refactor add Shape interface for area calculation
Defines a type `Shape` interface with an `Area()` return type float64 method.

Refactors area tests to use a helper function for improved readability and reduced code duplication.

We want to be able to write some kind of checkArea function that we can pass both Rectangles and Circles to, but fail to compile if we try to pass in something that isn't a shape.